extracted WebsiteScraper scraping logic into ReferenceScraper. Removed binaries. Wrote tests for ReferenceScraper.
I created a scraper interface with separate methods for each bit of parsing to be done. This will allow other scrapers to be written interchangably, and allow users to pick and choose which bits of info they want parsed (to save time, in the end).
I asked git to stop tracking binaries because there's no sense in tracking the files you're going to be building fresh anyway.
The same simple tests used for WebsiteScraper were 'ported' to ReferenceScraper.