Added service/receiver combo for phone-call alerts. Modified original GUI to be reusable. minor bug fixes.
Biggest change to original code: MinutesChecker is now called by `startActivityForResult`. This allows the code to return to _whichever_ activity called it (rather than hard-coded to ViewMinutes Activity). Necessary to reuse the login view/logic for the service. See `ViewMinutes.onActivityResult`.
Minor changes: moved some common logic into their own methods in the ViewMinutes class. See `showErrorMessageAndRequery`, `gatherAndDisplay`, and `setLoginInfo`.
New Service: Runs in the background and is automatically started the first time a phone call is received. It (currently) scrapes minute usage and displays a toast message on top of the call screen. This is slow over 3G, so I started down the path of using a persistent cache. I need to hook it up to update on GUI usage and to CALL_STATE_IDLE events.

I wouldn't call this commit stable.