cached-minutes displays on incoming call, updates on call finished. Fails semi-gracefully (annoyingly?)
On incoming call, the cache is queried for minutes data. If some is stored, that most recent cache data is displayed over the incoming message dialog (via Toast, default 10 seconds).
If cache data is not available, 2 things happen: 1) the message "Minutes Used: Unknown" is shown, and 2) when the call ends, the user is prompted to login to VM (via the original GUI), which will update the cache for them.

This could use a few configuration options, and some string resources instead of hardcoded strings. Configuration ideas:

 - whether to show minutes toast on incoming phone call?
 - whether to cache minutes on call ended? (showing min on incoming phone calls may be a little useless without this ... lump together, maybe?)

Potential bug: how quickly is VM's website updated? Should we wait 5+ minutes before caching?