Updated Virgin Mobile URL
I updated the virgin mobile URL. It required the ww2 instead of ww1.
(This is also true if logging in via a web broswer).

While running in the emulator it was not able to correctly pull the
charge date nor did it show the black line in the pie graph. I don't
know if this was somethign with my emulator, or something that will need
fixed.